refactor: move CompiledContractCache to vm_runner::cache module (#10817)
Based on top of https://github.com/near/nearcore/pull/10791
