+++ chain/client/src/test_utils/test_env_builder.rs
@@ -19,8 +19,7 @@ use near_primitives::types::{AccountId, NumShards};
 use near_store::config::StateSnapshotType;
 use near_store::test_utils::create_test_store;
 use near_store::{NodeStorage, ShardUId, Store, StoreConfig, TrieConfig};
-use near_vm_runner::logic::CompiledContractCache;
-use near_vm_runner::FilesystemCompiledContractCache;
+use near_vm_runner::{CompiledContractCache, FilesystemCompiledContractCache};
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::sync::Arc;
