+++ runtime/near-vm-runner/src/lib.rs
@@ -26,8 +26,8 @@ mod wasmtime_runner;
 
 pub use crate::logic::with_ext_cost_counter;
 pub use cache::{
-    get_contract_cache_key, precompile_contract, FilesystemCompiledContractCache,
-    MockCompiledContractCache,
+    get_contract_cache_key, precompile_contract, CompiledContract, CompiledContractCache,
+    FilesystemCompiledContractCache, MockCompiledContractCache,
 };
 pub use code::ContractCode;
 pub use profile::ProfileDataV2;
