+++ runtime/near-vm-runner/src/logic/mod.rs
@@ -1,6 +1,5 @@
 use borsh::{BorshDeserialize, BorshSerialize};
 use near_primitives_core::hash::CryptoHash;
-use std::fmt;
 use types::AccountId;
 
 mod alt_bn128;
@@ -25,40 +24,6 @@ pub use logic::{VMLogic, VMOutcome};
 pub use near_parameters::vm::{Config, ContractPrepareVersion, LimitConfig, StorageGetMode};
 pub use near_primitives_core::types::ProtocolVersion;
 pub use types::ReturnData;
-
-#[derive(Debug, Clone, PartialEq, BorshDeserialize, BorshSerialize)]
-pub enum CompiledContract {
-    CompileModuleError(errors::CompilationError),
-    Code(Vec<u8>),
-}
-
-impl CompiledContract {
-    /// Return the length of the compiled contract data.
-    ///
-    /// If the `CompiledContract` represents a compilation failure, returns `0`.
-    pub fn debug_len(&self) -> usize {
-        match self {
-            CompiledContract::CompileModuleError(_) => 0,
-            CompiledContract::Code(c) => c.len(),
-        }
-    }
-}
-
-/// Cache for compiled modules
-pub trait CompiledContractCache: Send + Sync {
-    fn handle(&self) -> Box<dyn CompiledContractCache>;
-    fn put(&self, key: &CryptoHash, value: CompiledContract) -> std::io::Result<()>;
-    fn get(&self, key: &CryptoHash) -> std::io::Result<Option<CompiledContract>>;
-    fn has(&self, key: &CryptoHash) -> std::io::Result<bool> {
-        self.get(key).map(|entry| entry.is_some())
-    }
-}
-
-impl fmt::Debug for dyn CompiledContractCache {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Compiled contracts cache")
-    }
-}
 
 /// Counts trie nodes reads during tx/receipt execution for proper storage costs charging.
 #[derive(Debug, PartialEq)]
@@ -95,40 +60,4 @@ impl TrieNodesCount {
 pub struct DataReceiver {
     pub data_id: CryptoHash,
     pub receiver_id: AccountId,
-}
-
-impl CompiledContractCache for Box<dyn CompiledContractCache> {
-    fn handle(&self) -> Box<dyn CompiledContractCache> {
-        <dyn CompiledContractCache>::handle(&**self)
-    }
-
-    fn put(&self, key: &CryptoHash, value: CompiledContract) -> std::io::Result<()> {
-        <dyn CompiledContractCache>::put(&**self, key, value)
-    }
-
-    fn get(&self, key: &CryptoHash) -> std::io::Result<Option<CompiledContract>> {
-        <dyn CompiledContractCache>::get(&**self, key)
-    }
-
-    fn has(&self, key: &CryptoHash) -> std::io::Result<bool> {
-        <dyn CompiledContractCache>::has(&**self, key)
-    }
-}
-
-impl<C: CompiledContractCache> CompiledContractCache for &C {
-    fn handle(&self) -> Box<dyn CompiledContractCache> {
-        <C as CompiledContractCache>::handle(self)
-    }
-
-    fn put(&self, key: &CryptoHash, value: CompiledContract) -> std::io::Result<()> {
-        <C as CompiledContractCache>::put(self, key, value)
-    }
-
-    fn get(&self, key: &CryptoHash) -> std::io::Result<Option<CompiledContract>> {
-        <C as CompiledContractCache>::get(self, key)
-    }
-
-    fn has(&self, key: &CryptoHash) -> std::io::Result<bool> {
-        <C as CompiledContractCache>::has(self, key)
-    }
 }
